requirements/base.txt
Django==5.2.*
djangorestframework>=3.16.1
django-filter==25.1

# Database
psycopg2-binary==2.9.10

# Authentication & Security
PyJWT==2.8.*
cryptography==41.*
django-oauth-toolkit==3.1.0
djangorestframework-simplejwt==5.5.*
django-allauth==65.11.2
django-cors-headers>=4.7.0
django-redis==6.0.*
redis==6.4.0
django-phonenumber-field[phonenumbers]==8.1.0

# API Frameworks
django-ninja==1.4.*

# API Documentation
drf-spectacular==0.27.*

# Background Tasks
celery==5.5.*

# File Storage
django-storages[s3]==1.14.*
boto3==1.40.*

# Internationalization
django-modeltranslation==0.19.*

# Utilities
python-decouple==3.8
Pillow>=10.3.0
python-dateutil==2.8.*
requests==2.31.*

scalar-django-ninja==0.1.4
